requirements.txt
fastapi
uvicorn[standard]
gunicorn
python-dotenv
requests
openai
exa-py
